chore: refactor account linking errors
