+++ packages/no-modal/src/account-linking/index.ts
@@ -1,2 +1,3 @@
+export * from "./errors";
 export * from "./interfaces";
 export * from "./rest";
